+++ uefi-raw/src/table/boot.rs
@@ -9,6 +9,7 @@ use crate::{
 };
 use bitflags::bitflags;
 use core::ffi::c_void;
+use core::mem::offset_of;
 use core::ops::RangeInclusive;
 
 newtype_enum! {
@@ -360,12 +361,15 @@ bitflags! {
 /// [version]: MemoryDescriptor::VERSION
 /// [0]: https://github.com/tianocore/edk2/blob/7142e648416ff5d3eac6c6d607874805f5de0ca8/MdeModulePkg/Core/PiSmmCore/Page.c#L1059
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[repr(C)]
+#[repr(C, align(8))]
 pub struct MemoryDescriptor {
     /// Type of memory occupying this range.
     pub ty: MemoryType,
-    /// Reserved field that must be set to 0.
-    pub reserved: u32,
+    /// Implicit padding that should be set to 0.
+    ///
+    /// This field is required for a correct layout on non-UEFI targets to
+    /// properly parse a memory map.
+    pub padding: u32,
     // Implicit 32-bit padding.
     /// Starting physical address.
     pub phys_start: PhysicalAddress,
@@ -377,6 +381,18 @@ pub struct MemoryDescriptor {
     pub att: MemoryAttribute,
 }
 
+// Ensure ABI guarantees for MemoryDescriptor.
+const _: () = {
+    assert!(size_of::<MemoryDescriptor>() == 40);
+    assert!(align_of::<MemoryDescriptor>() == 8);
+
+    assert!(offset_of!(MemoryDescriptor, ty) == 0);
+    assert!(offset_of!(MemoryDescriptor, phys_start) == 8);
+    assert!(offset_of!(MemoryDescriptor, virt_start) == 16);
+    assert!(offset_of!(MemoryDescriptor, page_count) == 24);
+    assert!(offset_of!(MemoryDescriptor, att) == 32);
+};
+
 impl MemoryDescriptor {
     /// Memory descriptor version number.
     pub const VERSION: u32 = 1;
@@ -386,7 +402,7 @@ impl Default for MemoryDescriptor {
     fn default() -> Self {
         Self {
             ty: MemoryType::RESERVED,
-            reserved: 0,
+            padding: 0,
             phys_start: 0,
             virt_start: 0,
             page_count: 0,