+++ uefi/tests/memory_map.rs
@@ -10,7 +10,7 @@ fn parse_boot_information_efi_mmap() {
     let mut mmap_source = [
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
-            reserved: 0,
+            padding: 0,
             phys_start: 0x3000,
             virt_start: 0x3000,
             page_count: 1,
@@ -18,7 +18,7 @@ fn parse_boot_information_efi_mmap() {
         },
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
-            reserved: 0,
+            padding: 0,
             phys_start: 0x2000,
             virt_start: 0x2000,
             page_count: 1,
@@ -26,7 +26,7 @@ fn parse_boot_information_efi_mmap() {
         },
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
-            reserved: 0,
+            padding: 0,
             phys_start: 0x1000,
             virt_start: 0x1000,
             page_count: 1,