+++ uefi-raw/src/lib.rs
@@ -105,6 +105,29 @@ impl Boolean {
 
     /// [`Boolean`] representing `false`.
     pub const FALSE: Self = Self(0);
+
+    /// Const-compatible check for **logical equality**.
+    const fn eq_const(self, other: Self) -> bool {
+        match (self.0, other.0) {
+            (0, 0) => true,
+            (0, _) => false,
+            (_, 0) => false,
+            // We handle it as in C: Any bit-pattern != 0 equals true
+            (_, _) => true,
+        }
+    }
+
+    /// Returns whether the underlying value equals a Rust `true`.
+    #[must_use]
+    pub const fn is_true(self) -> bool {
+        Self::eq_const(self, Self::TRUE)
+    }
+
+    /// Returns whether the underlying value equals a Rust `true`.
+    #[must_use]
+    pub const fn is_false(self) -> bool {
+        !self.is_true()
+    }
 }
 
 impl From<u8> for Boolean {
@@ -135,13 +158,7 @@ impl From<Boolean> for bool {
 
 impl PartialEq for Boolean {
     fn eq(&self, other: &Self) -> bool {
-        match (self.0, other.0) {
-            (0, 0) => true,
-            (0, _) => false,
-            (_, 0) => false,
-            // We handle it as in C: Any bit-pattern != 0 equals true
-            (_, _) => true,
-        }
+        Self::eq_const(*self, *other)
     }
 }