+++ uefi-raw/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Added
 
 ## Changed
+- **Breaking**: `MemoryDescriptor` now has a new member to ensure correct
+  padding on 32-bit targets.
 
 ## Removed